refactor container tests to separate files
